backend/requirements.txt
Django==4.2.11
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
Pillow==12.0.0
psycopg2-binary==2.9.11
django-filter==25.1
django-cors-headers==4.9.0
stripe==5.5.0
